requirements.txt
requests==2.21.0
spacy==2.0.16
pandas==0.23.4
numpy==1.15.4
Flask==1.0.2
python-dotenv==0.10.1
scikit_learn==0.20.2
gunicorn==19.9.0